refactor: Move application ports to core ports and update dependencies
