+++ core/ports/DocumentRepository.ts
